main/requirements.txt
flask==3.0.3
flask-cors==4.0.0
gunicorn==22.0.0
python-dotenv==1.0.1
openai==1.12.0
azure-identity==1.17.1
azure-ai-projects==1.0.0b7
azure-ai-agents==1.1.0
waitress==2.1.2